🌟 feat: diary cqrs 패턴 도입
