+++ src/main/java/com/x1/groo/diary/command/domain/dto/DiaryDraftRequestDTO.java
@@ -1,0 +1,17 @@
+package com.x1.groo.diary.command.domain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDate;
+
+@AllArgsConstructor
+@Getter
+@Setter
+public class DiaryDraftRequestDTO {
+
+    private LocalDate date;
+    private String content;
+    private int forestId;
+}
