+++ src/main/java/com/x1/groo/diary/command/domain/dto/DiaryUpdateRequestDTO.java
@@ -1,0 +1,10 @@
+package com.x1.groo.diary.command.domain.dto;
+
+import lombok.Data;
+
+@Data
+public class DiaryUpdateRequestDTO {
+    private String content;
+    private int forestId;
+    private int diaryId;
+}
