+++ src/main/java/com/x1/groo/diary/command/domain/dto/DiaryUpdateResponseDTO.java
@@ -1,0 +1,15 @@
+package com.x1.groo.diary.command.domain.dto;
+
+import lombok.Data;
+
+import java.time.LocalDateTime;
+import java.util.Date;
+
+@Data
+public class DiaryUpdateResponseDTO {
+    private int diaryId;
+    private int userId;
+    private int forestId;
+    private String content;
+    private LocalDateTime updatedAt;
+}
