+++ src/main/java/com/x1/groo/diary/command/domain/entity/DiaryEmotion.java
@@ -1,0 +1,24 @@
+package com.x1.groo.diary.command.domain.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "diary_emotion")
+@Getter @Setter @NoArgsConstructor
+public class DiaryEmotion {
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    @Column(nullable = false)
+    private int weight;
+
+    @Column(nullable = false)
+    private String emotion;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "diary_id", nullable = false)
+    private Diary diary;
+}
