--- src/main/java/com/x1/groo/diary/dto/DiaryDraftRequestDTO.java
@@ -1,16 +1,0 @@
-package com.x1.groo.diary.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-import java.time.LocalDate;
-
-@AllArgsConstructor
-@Getter
-@Setter
-public class DiaryDraftRequestDTO {
-
-    private LocalDate date;
-    private String content;
-}
