+++ src/main/java/com/x1/groo/diary/query/dto/ResponseDiaryDetailDTO.java
@@ -1,0 +1,18 @@
+package com.x1.groo.diary.query.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+@NoArgsConstructor
+@Getter
+@Setter
+public class ResponseDiaryDetailDTO {
+    private int diaryId;                // 일기 ID (감정 매핑용)
+    private String content;             // 일기의 내용
+    private LocalDateTime createdAt;    // 일기 작성 날짜 + 시간
+    private List<String> emotions;      // 감정 리스트
+}
