+++ src/main/java/com/x1/groo/diary/query/dto/ResponseSharedDiaryListDTO.java
@@ -1,0 +1,16 @@
+package com.x1.groo.diary.query.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@AllArgsConstructor
+@Getter
+@Setter
+public class ResponseSharedDiaryListDTO {
+    private int diaryId;
+    private int userId;
+    private LocalDateTime createdAt;
+}
