--- src/main/java/com/x1/groo/diary/repository/DiaryRepository.java
@@ -1,12 +1,0 @@
-package com.x1.groo.diary.repository;
-
-import com.x1.groo.diary.entity.Diary;
-import com.x1.groo.forest.common.domain.aggregate.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public interface DiaryRepository extends JpaRepository<Diary, Integer> {
-    void deleteAllByUserId(int userId);
-}
