+++ src/main/java/com/x1/groo/forest/emotion/query/repository/QueryForestEmotionMapper.java
@@ -6,12 +6,8 @@ import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionMailboxDTO;
 import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionMailboxListDTO;
 import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionUserItemDTO;
 import java.util.List;
-import com.x1.groo.forest.emotion.query.dto.*;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
-
-import java.time.LocalDateTime;
-import java.util.List;
 
 @Mapper
 public interface QueryForestEmotionMapper {
@@ -36,28 +32,6 @@ public interface QueryForestEmotionMapper {
     );
 
     List<QueryForestEmotionDetailDTO> findForestDetail(
-            @Param("userId") int userId,
-            @Param("forestId") int forestId
-    );
-
-    List<QueryForestEmotionDiaryByDateDTO> findDiaryByDateAndForestId(
-            @Param("forestId") int forestId,
-            @Param("startDateTime") LocalDateTime startDateTime,
-            @Param("endDateTime") LocalDateTime endDateTime
-    );
-
-    List<QueryForestEmotionDiaryByMonthDTO> findDiariesByMonth(
-            @Param("forestId") int forestId,
-            @Param("startDateTime") LocalDateTime startDateTime,
-            @Param("endDateTime") LocalDateTime endDateTime
-    );
-
-    Boolean existsUserInForest(
-            @Param("userId") int userId,
-            @Param("forestId") int forestId
-    );
-
-    Boolean isOwnerOfForest(
             @Param("userId") int userId,
             @Param("forestId") int forestId
     );
