+++ src/main/java/com/x1/groo/forest/emotion/query/service/QueryForestEmotionService.java
@@ -5,10 +5,8 @@ import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionListDTO;
 import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionMailboxDTO;
 import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionMailboxListDTO;
 import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionUserItemDTO;
-import com.x1.groo.forest.emotion.query.dto.*;
 import org.springframework.security.access.AccessDeniedException;
 
-import java.time.LocalDate;
 import java.util.List;
 
 public interface QueryForestEmotionService {
@@ -19,10 +17,6 @@ public interface QueryForestEmotionService {
     List<QueryForestEmotionMailboxDTO> getMailboxDetail(int userId, int id);
 
     List<QueryForestEmotionDetailDTO> getForestDetail(int userId, int forestId);
-
-    List<QueryForestEmotionDiaryByDateDTO> findDiaries(int userId, int forestId, LocalDate date);
-
-    List<QueryForestEmotionDiaryByMonthDTO> findDiariesByMonth(int userId, int forestId, int year, int month);
 
     List<QueryForestEmotionListDTO> getForestList(int userId);
 }
