+++ src/main/java/com/x1/groo/forest/emotion/query/service/QueryForestEmotionServiceImpl.java
@@ -7,13 +7,10 @@ import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionListDTO;
 import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionMailboxDTO;
 import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionMailboxListDTO;
 import com.x1.groo.forest.emotion.query.dto.QueryForestEmotionUserItemDTO;
-import com.x1.groo.forest.emotion.query.dto.*;
 import com.x1.groo.forest.emotion.query.repository.QueryForestEmotionMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -48,43 +45,6 @@ public class QueryForestEmotionServiceImpl implements QueryForestEmotionService 
             throw new CustomException(ErrorCode.FOREST_ACCESS_DENIED);
         }
         return result;
-    }
-
-    // 날짜별 일기 조회
-    @Override
-    public List<QueryForestEmotionDiaryByDateDTO> findDiaries(int userId, int forestId, LocalDate date) {
-
-        boolean isOwner = queryForestEmotionMapper.isOwnerOfForest(userId, forestId);
-        boolean isShared = queryForestEmotionMapper.existsUserInForest(userId, forestId);
-
-        if (!(isOwner || isShared)) {
-            throw new CustomException(ErrorCode.FOREST_ACCESS_DENIED);
-        }
-
-        // 날짜 범위 생성
-        LocalDateTime startDateTime = date.atStartOfDay();      // 00:00:00
-        LocalDateTime endDateTime = date.atTime(23, 59, 59);    // 23:59:59
-
-        // 일기 리스트 조회
-        return queryForestEmotionMapper.findDiaryByDateAndForestId(forestId, startDateTime, endDateTime);
-    }
-
-    // 월 별 일기 조회
-    @Override
-    public List<QueryForestEmotionDiaryByMonthDTO> findDiariesByMonth(int userId, int forestId, int year, int month) {
-        boolean isOwner = queryForestEmotionMapper.isOwnerOfForest(userId, forestId);
-        boolean isShared = queryForestEmotionMapper.existsUserInForest(userId, forestId);
-
-        if (!(isOwner || isShared)) {
-            throw new CustomException(ErrorCode.FOREST_ACCESS_DENIED);
-        }
-
-        LocalDateTime startDateTime = LocalDate.of(year, month, 1).atStartOfDay();
-        LocalDateTime endDateTime = startDateTime
-                .withDayOfMonth(startDateTime.toLocalDate().lengthOfMonth())
-                .withHour(23).withMinute(59).withSecond(59);
-
-        return queryForestEmotionMapper.findDiariesByMonth(forestId, startDateTime, endDateTime);
     }
 
     public List<QueryForestEmotionListDTO> getForestList(int userId) {
