+++ src/main/resources/com/x1/groo/diary/query/repository/QueryDiaryMapper.xml
@@ -1,0 +1,78 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.x1.groo.diary.query.repository.QueryDiaryMapper">
+
+    <select id="getDiariesByYearMonth">
+        SELECT
+               d.id AS diary_id
+             , d.created_at
+          FROM diary d
+         WHERE d.user_id = #{userId}
+           AND d.forest_id = #{forestId}
+           AND d.created_at BETWEEN #{startDateTime} AND #{endDateTime}
+    </select>
+
+    <resultMap id="diaryDetailMap" type="com.x1.groo.diary.query.dto.ResponseDiaryDetailDTO">
+        <id property="diaryId" column="diary_id"/>
+        <result property="content" column="content"/>
+        <result property="createdAt" column="created_at"/>
+        <collection property="emotions" ofType="string">
+            <result column="emotion_name" />
+        </collection>
+    </resultMap>
+    <select id="getDiaryDetail" resultMap="diaryDetailMap">
+        SELECT
+               d.id AS diary_id
+             , d.content
+             , d.created_at
+             , de.emotion AS emotion_name
+          FROM diary d
+          LEFT JOIN diary_emotion de ON d.id = de.diary_id
+         WHERE d.id = #{diaryId}
+           AND d.user_id = #{userId}
+    </select>
+
+    <select id="getSharedDiariesByYearMonth">
+        SELECT
+               d.id AS diary_id
+             , d.user_id
+             , d.created_at
+          FROM diary d
+          JOIN forest f ON d.forest_id = f.id
+          JOIN shared_forest sf ON f.id = sf.forest_id
+         WHERE f.id = #{forestId}
+           AND sf.user_id = #{userId}
+    </select>
+
+    <resultMap id="sharedDiaryDetailMap" type="com.x1.groo.diary.query.dto.ResponseSharedDiaryDetailDTO">
+        <id property="diaryId" column="diary_id"/>
+        <result property="content" column="content"/>
+        <result property="createdAt" column="created_at"/>
+        <result property="userId" column="user_id" />
+        <result property="nickname" column="nickname" />
+        <collection property="emotions" ofType="string">
+            <result column="emotion_name" />
+        </collection>
+    </resultMap>
+    <select id="getSharedDiariesDetail" resultMap="sharedDiaryDetailMap">
+        SELECT
+               d.id AS diary_id
+             , d.content
+             , d.created_at
+             , u.id AS user_id
+             , u.nickname
+             , de.emotion AS emotion_name
+          FROM diary d
+          JOIN user u ON d.user_id = u.id
+          JOIN forest f ON d.forest_id = f.id
+          JOIN shared_forest sf ON f.id = sf.forest_id
+          JOIN diary_emotion de ON d.id = de.diary_id
+         WHERE d.id IN
+        <foreach item="id" collection="diaryIds" open="(" separator="," close=")">
+            #{id}
+        </foreach>
+        AND sf.user_id = #{userId}
+    </select>
+</mapper>
