+++ src/main/resources/com/x1/groo/forest/emotion/query/repository/QueryForestEmotionMapper.xml
@@ -60,20 +60,6 @@
         <result property="userId" column="user_id"/>
     </resultMap>
 
-    <resultMap id="DiaryWithEmotionResultMap" type="com.x1.groo.forest.emotion.query.dto.QueryForestEmotionDiaryByDateDTO">
-        <id property="diaryId" column="diary_id"/>
-        <result property="content" column="content"/>
-        <result property="createdAt" column="created_at"/>
-        <collection property="emotions" ofType="string">
-            <result column="emotion_name" />
-        </collection>
-    </resultMap>
-
-    <resultMap id="DiaryByMonthResultMap" type="com.x1.groo.forest.emotion.query.dto.QueryForestEmotionDiaryByMonthDTO">
-        <id property="diaryId" column="id"/>
-        <result property="createdAt" column="created_at"/>
-    </resultMap>
-
     <select id="findUserIdByForestId" resultType="int">
         SELECT user_id
         FROM forest
@@ -164,43 +150,5 @@
                 WHERE sf.forest_id = f.id)
          ORDER BY f.id DESC
          LIMIT 1;
-    </select>
-
-    <select id="findDiaryByDateAndForestId" resultMap="DiaryWithEmotionResultMap">
-        SELECT
-               d.id AS diary_id
-             , d.content
-             , d.created_at
-             , de.emotion AS emotion_name
-          FROM diary d
-          JOIN user u ON d.user_id = u.id
-          JOIN forest f ON d.forest_id = f.id
-          LEFT JOIN shared_forest sf ON f.id = sf.forest_id
-          LEFT JOIN diary_emotion de ON d.id = de.diary_id
-         WHERE f.id = #{forestId}
-           AND d.created_at BETWEEN #{startDateTime} AND #{endDateTime}
-         ORDER BY d.id DESC
-    </select>
-
-    <select id="findDiariesByMonth" resultMap="DiaryByMonthResultMap">
-        SELECT
-            d.id
-             , d.created_at
-        FROM diary d
-        WHERE d.forest_id = #{forestId}
-          AND d.created_at BETWEEN #{startDateTime} AND #{endDateTime}
-    </select>
-
-    <select id="isOwnerOfForest" resultType="boolean">
-        SELECT CASE WHEN COUNT(*) > 0 THEN TRUE ELSE FALSE END
-        FROM forest
-        WHERE id = #{forestId} AND user_id = #{userId}
-    </select>
-
-    <select id="existsUserInForest" resultType="boolean">
-        SELECT CASE WHEN COUNT(*) > 0 THEN TRUE ELSE FALSE END
-        FROM shared_forest
-        WHERE user_id = #{userId}
-          AND forest_id = #{forestId}
     </select>
 </mapper>
